Another Card Game Problem
Chef is playing a card game with his friend Rick Sanchez. He recently won against Rick's grandson Morty; however, Rick is not as easy to beat. The rules of this game are as follows:
     1. The power of a positive integer is the sum of digits of that integer. For example, the power of 13 is 1+3=4.
     2. Chef and Rick receive randomly generated positive integers. For each player, let's call the integer he received final power.
     3. The goal of each player is to generate a positive integer such that its power (defined above) is equal to his final power.
     4. The player who generated the integer with fewer digits wins the game. If both have the same number of digits, then Rick cheats and wins the game.
You are given the final power of Chef PC and the final power of Rick PR. Assuming that both players play optimally, find the winner of the game and the number of digits of the integer he generates.