Divide Candies
There are N heaps of candies (numbered 1 through N); for each valid i, the i-th of them contains exactly iK candies. You aim to partition these heaps between you and your friend in such a way that each heap is given to exactly one of you without being split among you.
The distribution should be as fair as possible. Formally, let's denote the total number of candies you get by A and the total number of candies your friend gets by B; your goal is to minimise |A−B|.
Find the smallest possible value of |A−B| and assign the heaps to you and your friend in some way that minimises |A−B|. If there are multiple solutions, you may find any one of them.